requirements.txt
asgiref==3.7.2
Django==4.2.4
lxml==4.9.3
Pillow==10.0.0
PyPDF2==3.0.1
python-docx==0.8.11
reportlab==4.0.4
sqlparse==0.4.4
pdf2docx~=0.5.6